// app/(ui)/component-library/template/anchor-section/anchor.types.ts
import { LucideIcon } from "lucide-react";
import {IconType} from "react-icons";
import {
    BlockDataWrapper,
    SectionBlock
} from "@/app/(ui)/component-library/template/block-section/block-section-renderer";

export interface AnchorSection {
    id: string;
    label: string;
    icon: LucideIcon | IconType;
}

export interface PageSection extends AnchorSection {
    description?: string;
    cols: 1 | 2 | 3 | 4;
    blocks: SectionBlock[];
    gatewayFn?: (legislature: number) => Promise<Record<string, BlockDataWrapper>>;
    lazy?: boolean;
}